requirements.txt
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
Flask==2.3.3
Flask-Cors==4.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.3.7
zipp==3.16.2
